+++ plugins/web-factory/worker/src/test-env.d.ts
@@ -1,5 +1,0 @@
-import type { Env } from "./auth";
-
-declare module "cloudflare:test" {
-  interface ProvidedEnv extends Env {}
-}